+++ e2e/config/config.go
@@ -316,6 +316,7 @@ func (a PolicyAccounts) AsSlice() []Account {
 func (c Config) Validate() error {
 	if c.RPCs.Bitcoin.Params != Mainnet &&
 		c.RPCs.Bitcoin.Params != Testnet3 &&
+		c.RPCs.Bitcoin.Params != Signet &&
 		c.RPCs.Bitcoin.Params != Regnet {
 		return errors.New("invalid bitcoin params")
 	}
@@ -522,6 +523,7 @@ type BitcoinNetworkType string
 const (
 	Mainnet  BitcoinNetworkType = "mainnet"
 	Testnet3 BitcoinNetworkType = "testnet3"
+	Signet   BitcoinNetworkType = "signet"
 	Regnet   BitcoinNetworkType = "regnet"
 )
 
@@ -532,6 +534,8 @@ func (bnt BitcoinNetworkType) GetParams() (chaincfg.Params, error) {
 		return chaincfg.MainNetParams, nil
 	case Testnet3:
 		return chaincfg.TestNet3Params, nil
+	case Signet:
+		return chaincfg.SigNetParams, nil
 	case Regnet:
 		return chaincfg.RegressionNetParams, nil
 	default: